requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
charset-normalizer==3.1.0
discord==2.2.2
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycodestyle==2.10.0
python-dotenv==1.0.0
tomli==2.0.1
yarl==1.8.2
